fix: avoid self connections caused by TCP simultaneous open
The manual detection was used because Rust doesn't provide an interface
to bind connecting socket.

Another option would be to use a C library with Rust bindings.

The `ErrorKind::ConnectionRefused` error is used, because it's the error
that shows up when you try to connect to a port for which there are no
servers listening.

Fixes #195