initial messaging layer
An alternative approach could be:

    enum Event {
        Contacts(Contacts),
    }

    enum Message {
        External(Bytes),
        Internal(Event),
    }

But this would be 2 levels of indirection that could be leading to
more inefficient representation. Also, current approach saves typing.

This commits also gets rid of parse_contacts, as the serialization
happens on a central placce.

This support is "initial" because there room for improvement, like adding other
events (e.g. ping...) and find code that could be obsoleted by this approach.