increase the interval to stop reading
This change will decrease the chance we're not reading from socket when
a packet arrives (i.e. decrease the chance to a packet loss event
occurs).

It shouldn't be needed, but rust-utp isn't being very reliable. At some
point, it might be required to patch rust-utp to update the congestion
control algorithm and the strategy to when resend a packet.