adds a inactivity timeout independent from low-level building blocks
This is a mechanism to protect from unreliability of rust-utp.
Currently, timeout is 5min, but can be changed to better suit Routing
needs. Thanks to only high-level events being detected (and
`cbor::Decoder` taking ownership of our chance to intercept "less than a
whole message received" events like "a bunch of bytes received"), this
mechanism may not be optimal if the crust user is hoping to send bigger
messages.