fix/timer: collect handle for new timeout scheduled as a result of prev timeout 1) Each time routing sends something, we reset the heartbeat timer, cancelling the previous one and collecting the handle to the new one: https://github.com/maidsafe/crust/blob/6cf1acb405a1ff2e94ab48483d8ccff53a5c451c/src/main/active_connection.rs#L274  - this is fine 2) Each time there is a timeout because routing did not send anything, timeout fires, we send the heartbeat and schedule timer again https://github.com/maidsafe/crust/blob/6cf1acb405a1ff2e94ab48483d8ccff53a5c451c/src/main/active_connection.rs#L258  - there is a bug here that we do not collect the timer handle (due to `if let Err(e) = core.set_timeout(....) {` thus ignoring the `Ok(handle)` case)
If only (1) happened, i.e. routing always sent something before the heartbeat timeout, everything would work fine. However if routing sent a bunch of thing and then nothing for next 20 secs (heartbeat timeout) then timeout would fire, crust would send a heartbeat and reschedule a timeout (for future heartbeat). This is fine but because of bug in (2) we never collected the handle and thus we have a zombie timer `A0` which will always fire, purge itself, come to (2) and spawn `A0` again. So a zombie `A0` will always spawn `A0` in each timeout and this leads to one entry being permanent in timer-queue.

If routing again did a bunch of writes, it would cancel previous timers setting new ones, and now when it does nothing for 20 secs - which should be a normal scenario - the last timer it scheduled say `A1` would fire, get into (2) and now we have 2 permanent entries `A1 and A2` which have no handles and will keep firing in the background, purging themselves but creating 2 new ones when each timesout.

So if this process of routing writing a bunch of stuff and then resting for 20 secs repeats say 1000 times per connection, we will have 1000 zombie timers in the background. If routing table size is 65 then we have reached the max timer capacity. When routing does a next write, there will be that timer overflow error and crust would severe the connection to that peer. The 1000 zombie timers in the background for this active-connection will eventually all fire (correctly purging themselves), but now since the connection state is terminated, it will be a no-op (timer-handler no longer exists). They will thus not reschedule new timeouts and we suddenly have 1000 empty slots.

These slots will again gradually fill due to other existing active-connections and new active-connections in future. There will be a connection reset due to overflow and slots will be freed again.

This explains the why we see the errors in spurts and not continually after the first error occurs.