feat/example: add common utilities for examples
This covers node ID implementation and connect_to() method which connects to a peer node. This code is so common that every example is expected to have it.
Note that I itentionally do not plan to refactor  example. I'd like it to be standalone because I believe that's easier to read for newcomers.