fix/connect: readd 'choose connection' logic
It was removed because identical code was in p2p crate. But as it turns out, this logic is not executed for direct connections.
This resulted in premature connections drops.
There are no semantic changes compared to before #887 was merged. It's just a copy paste.
The main idea is to return the right connection: one peer must return listening connection
while other peer must return connecting connection. Otherwise, connection loss would appear.