fix/bind_public: update p2p
bind_public() functions try to determine if we have public address. There was a bug in p2p that falsely returned public IP when we didn't actually have it. This prevented us from doing hole punched connections because if Crust detects that we have public IP, it doesn't event try hole punching.