fix/connect_all: fix direct incoming connection handling
We used to wait for incoming connections until all our outgoing
connections are done. Otherwise we would forever wait for incoming
connections and connect_all() would be blocked. But there was another
issue with this algorithm:

1. say peer1 does not have any public listeners, peer2 does,
2. then peer1 tries to connect with peer2
3. peer2 connect_all() connection attempts finish immediately, because
   `for_direct` was empty for peer1. Then incoming connections are
   terminated.
4. Then peer1 attempts direct connections, but peer2 is not listening
   for incoming connections anymore.
5. connection FAILS

This commit fixes that and makes sure connect_all() alway waits for
incoming connections at least for some time.