chore/ci: provide Dockerfile for build
Introduces a container for performing the build and test steps.

The `RUN` instruction will execute `cargo build` during the
`docker build` phase, and this will download and compile all the
dependencies in advance, baking them into the container for when it's
subsequently used during a CI process or locally.

The `CMD` instruction will be the default operation performed on a
`docker run`, making use of the pre-built dependencies.

Both these instructions tell Cargo to explicitly use an output
directory that's different from the default current directory. This is
because when the container is later used, the directory with the current
code will be mounted into the container with the `-v` option to
`/usr/src/crust`, meaning the default target directory would be
overwritten.