chore/ci: remove container pull targets
There isn't much point in having the utility target for pulling the
container because in the context of using Docker to build on a
development host, you want to run the build with the current code, and
the container you pull will potentially have code that's out of date.
The documentation is also updated to reflect this.