chore/ci: optimise container size
This provides a few optimisations to reduce the size of the container
from ~2.2GB to ~1.85GB:

* Use a .dockerignore to preclude the .git directory from being added to
the container
* Remove some space from the apt cache
* Remove any target directory before running a `docker build`
* Combine the `cargo build` command into the same layer as the
apt-update (reducing the number of layers reduces container size)

The Rust base image consumes 1.5GB, so as it stands I don't think it
would be possible to make this any smaller. If we needed to we could
look into using our own very minimal base image.