resolves #1077: implement and populate bootstrap cache (#1092)
* feat/bootstrap: implement bootstrap cache

Implemented in-memory bootstrap cache structure with possibility to sync data on disk.

* docs/event_loop: add some docs about event loop related structures

* feat/misc: update bootstrap cache during connection attempts

When a successful connection is made, the peer's contacts are cached so
that we could bootstrap off it in the future. If connection failed, it's
details are removed from the cache, if it previously existed there.

* feat/core: implement user data

Templatize Core to enable storing arbitrary user data.
Currently the main motivation is to store bootstrap cache which is meant
to only be used within a single thread. Therefore Core works great for
this since it lives on the same thread as our event loop.

* test/connections: test if bootstrap cache is updated properly

* chore/misc: update rust to 2018 edition

These changes include

* cargo fix --edition
* cargo fmt changes
* and clippy fixes

* feat/service: add method to retrieve bootstrap cache