feat/conn_listener: use GetExtAddr to test external reachability
Before external reachability test was just a simple TCP connectivity test. That didn't ensure
that tested address is actually a valid Crust peer.
Now this is changed and an encrypted Crust message is sent to remote peer then only a valid peer
with appropriate private key can respond to this message and only then it is considered to be
publicly reachable.

Also, I reused get GetExtAddr code for this to avoid introducing extra message. GetExtAddr
sends one encrypted request and waits for a single encrypted response. So, it doesn't add
any extra overhead and reduces code at the same time :)