refactor/messages: split Connect message into request/response
Before Crust used to use a single message to initiate connection and respond to such connection requests.
This lead to issues where response was carrying information that was not used at all. So segregating
request/response messages is a better design and makes the code cleaner and easier to follow.