refactor/misc: store listener addrs in Core
This effectively eliminates the need of Mutex for our listener
addresses which are always used from the same thread as
our event loop is running on.

As a side effect, service discovery module was moved into main
module because it had to be aware of user data type stored in Core in
order to be able to use our listener addresses.