resolves #1086: configurable external reachability test (#1104)
* docs/misc: expand service docs and add some private docs

* feat/conn_listener: make conn listener control ext reachability test

Before it was the bootstrapee who proactively asked to be tested
for external reachability. This commit changes the semantics and
makes connection listener to always test external reachability
for `Node` peers unless explicitly disabled.

* feat/misc: implement external reachability test during Service::connect()

Service::connect() is usually used for CrustUser::Node roles.
Therefore, we want to always test if peers are accessible direclty.
Hence external reachability test is executed during peer connection.
Although, this test can be disabled with
`Service::set_ext_reachability_test(false)`.

* feat/conn_listener: don't check IPs for external reachability that don't belong to connected peer

* refactor/config: remove redundant dev config

External reachability test is configurable from  so this option became redundant.

* refactor/misc: refactor naming

* feat/conn_listener: use GetExtAddr to test external reachability

Before external reachability test was just a simple TCP connectivity test. That didn't ensure
that tested address is actually a valid Crust peer.
Now this is changed and an encrypted Crust message is sent to remote peer then only a valid peer
with appropriate private key can respond to this message and only then it is considered to be
publicly reachable.

Also, I reused get GetExtAddr code for this to avoid introducing extra message. GetExtAddr
sends one encrypted request and waits for a single encrypted response. So, it doesn't add
any extra overhead and reduces code at the same time :)

* feat/conn_listener: limit max external reachability test addresses

* feat/conn_listener: elevate logging level

During debugging it might be useful to know why connection fails. Since
external reachability test is executed by default for Node peers, we
might bump into this issue, so better use debug log messages in those
cases.

* refactor/bootstrap: make our global addresses optional

If the peer with a role `Client` is bootstrapping, it is not required
to send it's global address list because we don't test external
reachability for `Client`s.