Split connect message (#1108)
* refactor/messages: split Connect message into request/response

Before Crust used to use a single message to initiate connection and respond to such connection requests.
This lead to issues where response was carrying information that was not used at all. So segregating
request/response messages is a better design and makes the code cleaner and easier to follow.

* refactor/types: move more type definitions into types.rs module

Makes it uniform with other types.