resolves #1089: reduce the use of mutexes (#1111)
* feat/event_loop: change user data type stored in Core

Created a new structure `CrustData` that wraps bootstrap cache
and will allow to store more data.

* refactor/misc: store listener addrs in Core

This effectively eliminates the need of Mutex for our listener
addresses which are always used from the same thread as
our event loop is running on.

* refactor/misc: store connection map in Core

That allows us to eliminate the use of one more mutex.

* refactor/misc: store Config in Core

Eliminates the need of mutex for this.

* refactor/crust_user: implement BootstrapperRole to CrustUser conversion trait

Makes code more Rustish.

* refactor/types: use HashSet to hold a list of peers

HashSet implies unique set. Hence it's more appropriate to store peer
list in it because we don't want to have duplicate peers in our lists.

* refactor/service_discovery: decouple it from main Crust code

To do that we introduced a special trait that is implemented by Crust
user data stored in Core, and ServDisc uses that to get the listener addresses to send to others asking.