resolves #1106: bootstrap cache improvements (#1120)
* refactor/bootstrap_cache: remove redundant Rc

Bootstrap cache is not shared anymore, rather it's stored in Core
whose lifetime is the same as event loop thread lifetime.

* refactor/misc: update docs, reformat imports

* feat/bootstrap: don't shuffle cached peers

We will be using the most recently active peers and bootstrapping off
them. This way we expect the highest success rate and shuffling
those peers does not buy as anything anymore.

* feat/bootstrap_cache: add limits to bootstrap cache

Added some sane default limits for the peers Crust will cache and use
during bootstrapping.  Use LRU cache to store cached peers so that we
would be using only the most recent entries.

* feat/bootstrap_cache: test expired peers

1. When cached bootstrap peers become inactive,
   they will be queued for liveness test.
2. Then Crust will send STUN request to every such peer.
3. Peers that respond to this request are considered alive
   and are readded to the bootstrap cache.
4. Peers that don't respond are discarded.

* chore/deps: update lru_time_cache

This crate will be updated again once new version is released.

* fix/connect: test expired bootstrap cache peers

Forgot to reenable testing expired peers in this specific scenario.