feat/bootstrap_cache: don't explicitly remove peers from the cache
If we fail to connect to some peer, it does not mean we won't be able to
connect to it in 5 minutes or so.  Hence, don't just remove cached
peers. Peers will be evicted from the bootstrap cache, if there's no
more room for new directly accessible peers.

Keep remove() function for future use which will definitely happen.